requirements.txt
# Core runtime
pydantic>=2.7,<3.0
PyYAML>=6.0,<7.0

# HTTP + resilience (optional for web fetching)
httpx>=0.27,<0.29
tenacity>=8.2,<9.0

# Logging / observability
structlog>=24.1,<25.0
python-json-logger>=2.0,<3.0

# Data handling
pandas>=2.2,<3.0

# Timezone-safe timestamps
python-dateutil>=2.9,<3.0

#UI
streamlit>=1.37,<2.0

PyYAML>=6.0,<7.0


